# 1-foundation/5-database/1-sqlpuro/main.go
package main

import (
	"database/sql"
	"fmt"
	_ "github.com/go-sql-driver/mysql"
	"github.com/google/uuid"
)

type Product struct {
	ID    string
	Name  string
	Price float64
}

func NewProduct(name string, price float64) *Product {
	return &Product{
		ID:    uuid.New().String(),
		Name:  name,
		Price: price,
	}
}

func main() {
	db, err := sql.Open("mysql", "root:root@tcp(localhost:3306)/goexpert")
	if err != nil {
		panic(err)
	}
	defer db.Close()

	// CRIANDO UM PRODUTO
	//product := NewProduct("Book", 23.19)
	//err = insertProduct(db, product)
	//if err != nil {
	//	panic(err)
	//}
	//fmt.Println("Produto criado com sucesso")

	// ATUALIZANDO UM PRODUTO e5b294eb-cc72-42a0-aefa-4335443d8591
	//tablet := NewProduct("Tablet", 500)
	//tablet.ID = "e5b294eb-cc72-42a0-aefa-4335443d8591"
	//err = updateProduct(db, tablet)
	//if err != nil {
	//	panic(err)
	//}
	//fmt.Println("Produto atualizado com sucesso")

	// CONSULTANDO PRODUTO POR ID
	//p, err := selectProduct(db, "e5b294eb-cc72-42a0-aefa-4335443d8591")
	//if err != nil {
	//	panic(err)
	//}
	//fmt.Printf("Produto encontrado name= %s, price= %.2f", p.Name, p.Price)

	// CONSULTANDO TODOS OS PRODUTOS
	//products, err := selectAllProducts(db)
	//if err != nil {
	//	panic(err)
	//}
	//for _, p := range products {
	//	fmt.Printf("Produto name= %s, price= %.2f\n", p.Name, p.Price)
	//}

	// DELETANDO REGISTROS
	err = deleteProduct(db, "e5b294eb-cc72-42a0-aefa-4335443d8591")
	if err != nil {
		panic(err)
	}
	fmt.Println("Produto deletado com sucesso")
}

func insertProduct(db *sql.DB, product *Product) error {
	stmt, err := db.Prepare("INSERT INTO products(id, name, price) VALUES(?, ?, ?)")
	if err != nil {
		return err
	}
	defer stmt.Close()

	_, err = stmt.Exec(product.ID, product.Name, product.Price)
	if err != nil {
		return err
	}

	return nil
}

func updateProduct(db *sql.DB, product *Product) error {
	stmt, err := db.Prepare("UPDATE products SET name = ?, price = ? WHERE id = ?")
	if err != nil {
		return err
	}
	defer stmt.Close()

	_, err = stmt.Exec(product.Name, product.Price, product.ID)
	if err != nil {
		return err
	}

	return nil
}

func selectProduct(db *sql.DB, id string) (*Product, error) {
	stmt, err := db.Prepare("SELECT id, name, price FROM products WHERE id = ?")
	if err != nil {
		return nil, err
	}
	defer stmt.Close()

	var p Product
	err = stmt.QueryRow(id).Scan(&p.ID, &p.Name, &p.Price)
	// exemplo utilizando context para configurar timeout, e etc
	// err = stmt.QueryRowContext(ctx, id).Scan(&p.ID, &p.Name, &p.Price)
	if err != nil {
		return nil, err
	}

	return &p, nil
}

func selectAllProducts(db *sql.DB) ([]*Product, error) {
	rows, err := db.Query("SELECT id, name, price FROM products")
	if err != nil {
		return nil, err
	}
	defer rows.Close()

	var products []*Product
	for rows.Next() {
		var p Product
		err = rows.Scan(&p.ID, &p.Name, &p.Price)
		if err != nil {
			return nil, err
		}

		products = append(products, &p)
	}

	return products, nil
}

func deleteProduct(db *sql.DB, id string) error {
	stmt, err := db.Prepare("DELETE FROM products WHERE id = ?")
	if err != nil {
		return err
	}
	defer stmt.Close()

	_, err = stmt.Exec(id)
	if err != nil {
		return err
	}

	return nil
}